dashboard_foundation.py
#!/usr/bin/env python3
"""
Dashboard Foundation for Future Jarvis AI Integration
Minimal Flask foundation - implement later when monetization is active
"""

from flask import Flask, render_template, jsonify
import sqlite3

app = Flask(__name__)

class DashboardFoundation:
    def __init__(self):
        self.db_path = "amazon_intelligence.db"
    
    def get_dashboard_stats(self):
        """Basic stats for future dashboard"""
        conn = sqlite3.connect(self.db_path)
        cursor = conn.cursor()
        
        stats = {}
        
        # Product counts by status
        cursor.execute('''
            SELECT status, COUNT(*) 
            FROM product_opportunities 
            GROUP BY status
        ''')
        stats['status_counts'] = dict(cursor.fetchall())
        
        # Network coverage
        cursor.execute('''
            SELECT COUNT(DISTINCT p.id), COUNT(al.id)
            FROM product_opportunities p
            LEFT JOIN affiliate_links al ON p.id = al.product_id
        ''')
        total_products, total_links = cursor.fetchone()
        stats['network_coverage'] = {
            'total_products': total_products,
            'total_links': total_links,
            'avg_networks_per_product': round(total_links / total_products, 2) if total_products > 0 else 0
        }
        
        conn.close()
        return stats

# Basic API endpoints for future Jarvis integration
@app.route('/api/stats')
def api_stats():
    foundation = DashboardFoundation()
    return jsonify(foundation.get_dashboard_stats())

@app.route('/api/products')
def api_products():
    """Future endpoint for Jarvis AI integration"""
    return jsonify({"message": "Dashboard foundation ready for Jarvis AI integration"})

if __name__ == "__main__":
    print("🔧 Dashboard Foundation Ready")
    print("This is a minimal foundation for future Jarvis AI dashboard")
    print("Focus on monetization first - implement full dashboard later")
    
    # Don't start server automatically - just foundation code
    # app.run(debug=True, port=5000)