requirements.txt
openai
python-dotenv
requests
beautifulsoup4

